Spring cloud / refactoring common dependencies, jhipter registry docker
