+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/common/domain/SpringCloudCommonService.java
@@ -3,5 +3,9 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.common.domain
 import tech.jhipster.lite.generator.project.domain.Project;
 
 public interface SpringCloudCommonService {
+  void addSpringCloudCommonDependencies(Project project);
+
+  void addJhipsterRegistryDockerCompose(Project project);
+
   void addOrMergeBootstrapProperties(Project project, String sourceFolderPath, String sourceFileName, String destinationFileFolder);
 }
