+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/common/infrastructure/SpringCloudCommonBeanConfiguration.java
@@ -2,6 +2,7 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.common.infras
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonDomainService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
@@ -10,13 +11,15 @@ import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.
 public class SpringCloudCommonBeanConfiguration {
 
   private final ProjectRepository projectRepository;
+  private final BuildToolService buildToolService;
 
-  public SpringCloudCommonBeanConfiguration(ProjectRepository projectRepository) {
+  public SpringCloudCommonBeanConfiguration(ProjectRepository projectRepository, BuildToolService buildToolService) {
     this.projectRepository = projectRepository;
+    this.buildToolService = buildToolService;
   }
 
   @Bean
   public SpringCloudCommonService springCloudCommonService() {
-    return new SpringCloudCommonDomainService(projectRepository);
+    return new SpringCloudCommonDomainService(projectRepository, buildToolService);
   }
 }
