+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/configclient/domain/SpringCloudConfigClientDomainService.java
@@ -3,29 +3,21 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.configclient.
 import static tech.jhipster.lite.common.domain.FileUtils.getPath;
 import static tech.jhipster.lite.generator.project.domain.Constants.*;
 import static tech.jhipster.lite.generator.project.domain.DefaultConfig.BASE_NAME;
-import static tech.jhipster.lite.generator.server.springboot.springcloud.configclient.domain.SpringCloudConfig.*;
+import static tech.jhipster.lite.generator.server.springboot.springcloud.configclient.domain.SpringCloudConfig.springCloudConfigClient;
 
-import tech.jhipster.lite.common.domain.Base64Utils;
 import tech.jhipster.lite.error.domain.GeneratorException;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.project.domain.Project;
-import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
 
 public class SpringCloudConfigClientDomainService implements SpringCloudConfigClientService {
 
   private static final String SOURCE = "server/springboot/springcloud/configclient";
 
-  private final ProjectRepository projectRepository;
   private final BuildToolService buildToolService;
   private final SpringCloudCommonService springCloudCommonService;
 
-  public SpringCloudConfigClientDomainService(
-    ProjectRepository projectRepository,
-    BuildToolService buildToolService,
-    SpringCloudCommonService springCloudCommonService
-  ) {
-    this.projectRepository = projectRepository;
+  public SpringCloudConfigClientDomainService(BuildToolService buildToolService, SpringCloudCommonService springCloudCommonService) {
     this.buildToolService = buildToolService;
     this.springCloudCommonService = springCloudCommonService;
   }
@@ -39,18 +31,7 @@ public class SpringCloudConfigClientDomainService implements SpringCloudConfigCl
 
   @Override
   public void addDockerCompose(Project project) {
-    project.addDefaultConfig(BASE_NAME);
-    project.addConfig("jhipsterRegistryDockerImage", SpringCloudConfig.getJhipsterRegistryDockerImage());
-    projectRepository.template(project, SOURCE, "jhipster-registry.yml", "src/main/docker", "jhipster-registry.yml");
-
-    project.addConfig("base64JwtSecret", Base64Utils.getBase64Secret());
-    projectRepository.template(
-      project,
-      SOURCE,
-      "application.config.properties",
-      "src/main/docker/central-server-config/localhost-config",
-      "application.properties"
-    );
+    springCloudCommonService.addJhipsterRegistryDockerCompose(project);
   }
 
   @Override
@@ -82,9 +63,7 @@ public class SpringCloudConfigClientDomainService implements SpringCloudConfigCl
     this.buildToolService.getVersion(project, "spring-cloud")
       .ifPresentOrElse(
         version -> {
-          this.buildToolService.addProperty(project, "spring-cloud.version", version);
-          this.buildToolService.addDependencyManagement(project, springCloudDependencies());
-          this.buildToolService.addDependency(project, springCloudBootstrap());
+          this.springCloudCommonService.addSpringCloudCommonDependencies(project);
           this.buildToolService.addDependency(project, springCloudConfigClient());
         },
         () -> {
