+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/configclient/infrastructure/config/SpringCloudConfigClientBeanConfiguration.java
@@ -3,7 +3,6 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.configclient.
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.configclient.domain.SpringCloudConfigClientDomainService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.configclient.domain.SpringCloudConfigClientService;
@@ -11,22 +10,16 @@ import tech.jhipster.lite.generator.server.springboot.springcloud.configclient.d
 @Configuration
 public class SpringCloudConfigClientBeanConfiguration {
 
-  private final ProjectRepository projectRepository;
   private final BuildToolService buildToolService;
   private final SpringCloudCommonService springCloudCommonService;
 
-  public SpringCloudConfigClientBeanConfiguration(
-    ProjectRepository projectRepository,
-    BuildToolService buildToolService,
-    SpringCloudCommonService springCloudCommonService
-  ) {
-    this.projectRepository = projectRepository;
+  public SpringCloudConfigClientBeanConfiguration(BuildToolService buildToolService, SpringCloudCommonService springCloudCommonService) {
     this.buildToolService = buildToolService;
     this.springCloudCommonService = springCloudCommonService;
   }
 
   @Bean
   public SpringCloudConfigClientService springCloudConfigClientService() {
-    return new SpringCloudConfigClientDomainService(projectRepository, buildToolService, springCloudCommonService);
+    return new SpringCloudConfigClientDomainService(buildToolService, springCloudCommonService);
   }
 }
