+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/consul/domain/ConsulDomainService.java
@@ -42,9 +42,7 @@ public class ConsulDomainService implements ConsulService {
     this.buildToolService.getVersion(project, "spring-cloud")
       .ifPresentOrElse(
         version -> {
-          buildToolService.addProperty(project, "spring-cloud.version", version);
-          buildToolService.addDependencyManagement(project, springCloudDependencyManagement());
-          buildToolService.addDependency(project, springCloudBootstrapDependency());
+          springCloudCommonService.addSpringCloudCommonDependencies(project);
           buildToolService.addDependency(project, springCloudConsulConfigDependency());
           buildToolService.addDependency(project, springCloudConsulDiscoveryDependency());
         },
