+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/eureka/infrastructure/config/EurekaBeanConfiguration.java
@@ -3,7 +3,6 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.eureka.infras
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.eureka.domain.EurekaDomainService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.eureka.domain.EurekaService;
@@ -12,21 +11,15 @@ import tech.jhipster.lite.generator.server.springboot.springcloud.eureka.domain.
 public class EurekaBeanConfiguration {
 
   private final BuildToolService buildToolService;
-  private final ProjectRepository projectRepository;
   private final SpringCloudCommonService springCloudCommonService;
 
-  public EurekaBeanConfiguration(
-    BuildToolService buildToolService,
-    ProjectRepository projectRepository,
-    SpringCloudCommonService springCloudCommonService
-  ) {
+  public EurekaBeanConfiguration(BuildToolService buildToolService, SpringCloudCommonService springCloudCommonService) {
     this.buildToolService = buildToolService;
-    this.projectRepository = projectRepository;
     this.springCloudCommonService = springCloudCommonService;
   }
 
   @Bean
   public EurekaService eurekaService() {
-    return new EurekaDomainService(buildToolService, projectRepository, springCloudCommonService);
+    return new EurekaDomainService(buildToolService, springCloudCommonService);
   }
 }
