+++ src/test/java/tech/jhipster/lite/generator/server/springboot/springcloud/common/domain/SpringCloudCommonDomainServiceTest.java
@@ -1,22 +1,29 @@
 package tech.jhipster.lite.generator.server.springboot.springcloud.common.domain;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
+import static tech.jhipster.lite.TestUtils.tmpProject;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import tech.jhipster.lite.UnitTest;
+import tech.jhipster.lite.common.domain.Base64Utils;
 import tech.jhipster.lite.common.domain.FileUtils;
 import tech.jhipster.lite.error.domain.GeneratorException;
+import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
+import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 
@@ -31,8 +38,104 @@ class SpringCloudCommonDomainServiceTest {
   @Mock
   ProjectRepository projectRepository;
 
+  @Mock
+  BuildToolService buildToolService;
+
   @InjectMocks
   SpringCloudCommonDomainService springCloudCommonDomainService;
+
+  private Dependency getExpectedManagementDependency() {
+    return Dependency
+      .builder()
+      .groupId("org.springframework.cloud")
+      .artifactId("spring-cloud-dependencies")
+      .version("\\${spring-cloud.version}")
+      .type("pom")
+      .scope("import")
+      .build();
+  }
+
+  private Dependency getExpectedStarterBootstrapDependency() {
+    return Dependency.builder().groupId("org.springframework.cloud").artifactId("spring-cloud-starter-bootstrap").build();
+  }
+
+  @Nested
+  class AddSpringCloudCommonDependenciesTest {
+
+    @Test
+    void shouldAddSpringCloudCommonDependencies() {
+      // Given
+      Project project = tmpProject();
+      when(buildToolService.getVersion(project, "spring-cloud")).thenReturn(Optional.of("0.0.0"));
+
+      // When
+      springCloudCommonDomainService.addSpringCloudCommonDependencies(project);
+
+      // Then
+      verify(buildToolService).addProperty(project, "spring-cloud.version", "0.0.0");
+
+      ArgumentCaptor<Dependency> dependencyArgCaptor = ArgumentCaptor.forClass(Dependency.class);
+      verify(buildToolService).addDependencyManagement(eq(project), dependencyArgCaptor.capture());
+      assertThat(dependencyArgCaptor.getValue()).usingRecursiveComparison().isEqualTo(getExpectedManagementDependency());
+
+      verify(buildToolService).addDependency(eq(project), dependencyArgCaptor.capture());
+      assertThat(dependencyArgCaptor.getValue()).usingRecursiveComparison().isEqualTo(getExpectedStarterBootstrapDependency());
+    }
+
+    @Test
+    void shouldThrowExceptionWhenSpringCloudDependencyVersionNotFound() {
+      // Given
+      Project project = tmpProject();
+
+      when(buildToolService.getVersion(project, "spring-cloud")).thenReturn(Optional.empty());
+
+      // When + Then
+      assertThatThrownBy(() -> springCloudCommonDomainService.addSpringCloudCommonDependencies(project))
+        .isExactlyInstanceOf(GeneratorException.class);
+    }
+  }
+
+  @Nested
+  class AddJhipsterRegistryDockerComposeTest {
+
+    @Test
+    void shouldAddJhipsterRegistryDockerCompose() {
+      // Given
+      Project project = tmpProject();
+
+      String expectedBase64Secret = "encodedSecret";
+
+      try (MockedStatic<Base64Utils> base64Utils = mockStatic(Base64Utils.class)) {
+        base64Utils.when(Base64Utils::getBase64Secret).thenReturn(expectedBase64Secret);
+
+        // When
+        springCloudCommonDomainService.addJhipsterRegistryDockerCompose(project);
+
+        // Then
+        assertThat(project.getConfig("jhipsterRegistryDockerImage")).contains("jhipster/jhipster-registry:v7.1.0");
+        assertThat(project.getConfig("base64JwtSecret")).contains(expectedBase64Secret);
+
+        verify(projectRepository, times(2)).template(any(Project.class), anyString(), anyString(), anyString(), anyString());
+        verify(projectRepository)
+          .template(
+            project,
+            "server/springboot/springcloud/configclient",
+            "jhipster-registry.yml",
+            "src/main/docker",
+            "jhipster-registry.yml"
+          );
+
+        verify(projectRepository)
+          .template(
+            project,
+            "server/springboot/springcloud/configclient",
+            "application.config.properties",
+            "src/main/docker/central-server-config/localhost-config",
+            "application.properties"
+          );
+      }
+    }
+  }
 
   @Nested
   class AddOrMergeBootstrapPropertiesTest {
@@ -91,12 +194,12 @@ class SpringCloudCommonDomainServiceTest {
           .when(() -> FileUtils.read(destinationFilePath))
           .thenReturn(
             """
-            # another comment
-            prop1=existingValueEditedByUser
-            prop2=value2
-            prop4=valueAddedByUser
+              # another comment
+              prop1=existingValueEditedByUser
+              prop2=value2
+              prop4=valueAddedByUser
 
-            """
+              """
           );
 
         // When
@@ -127,9 +230,9 @@ class SpringCloudCommonDomainServiceTest {
 
         when(projectRepository.getComputedTemplate(project, SOURCE_FOLDER_PATH, SOURCE_FILE_NAME))
           .thenReturn("""
-              # a comment
-              prop1=valueFromGenerator1
-              """);
+            # a comment
+            prop1=valueFromGenerator1
+            """);
 
         fileUtils
           .when(() -> FileUtils.read(destinationFilePath))
