+++ src/test/java/tech/jhipster/lite/generator/server/springboot/springcloud/configclient/domain/SpringCloudConfigClientDomainServiceTest.java
@@ -17,15 +17,11 @@ import tech.jhipster.lite.error.domain.GeneratorException;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
 import tech.jhipster.lite.generator.project.domain.Project;
-import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
 
 @UnitTest
 @ExtendWith(MockitoExtension.class)
 class SpringCloudConfigClientDomainServiceTest {
-
-  @Mock
-  ProjectRepository projectRepository;
 
   @Mock
   BuildToolService buildToolService;
@@ -43,12 +39,11 @@ class SpringCloudConfigClientDomainServiceTest {
 
     springCloudConfigClientDomainService.init(project);
 
-    verify(buildToolService).addProperty(any(Project.class), anyString(), anyString());
-    verify(buildToolService).addDependencyManagement(any(Project.class), any(Dependency.class));
-    verify(buildToolService, times(2)).addDependency(any(Project.class), any(Dependency.class));
+    verify(springCloudCommonService).addSpringCloudCommonDependencies(project);
+    verify(buildToolService, times(1)).addDependency(any(Project.class), any(Dependency.class));
 
     verify(springCloudCommonService, times(3)).addOrMergeBootstrapProperties(any(Project.class), anyString(), anyString(), anyString());
-    verify(projectRepository, times(2)).template(any(Project.class), anyString(), anyString(), anyString(), anyString());
+    verify(springCloudCommonService).addJhipsterRegistryDockerCompose(any(Project.class));
   }
 
   @Test
