+++ src/test/java/tech/jhipster/lite/generator/server/springboot/springcloud/configclient/domain/SpringCloudConfigTest.java
@@ -10,27 +10,6 @@ import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
 class SpringCloudConfigTest {
 
   @Test
-  void shouldGetJhipsterRegistryDockerImageName() {
-    assertThat(SpringCloudConfig.getJhipsterRegistryDockerImage()).isEqualTo("jhipster/jhipster-registry:v7.1.0");
-  }
-
-  @Test
-  void shouldGetSpringCloudBootstrap() {
-    Dependency dependency = SpringCloudConfig.springCloudBootstrap();
-
-    assertThat(dependency.getGroupId()).isEqualTo("org.springframework.cloud");
-    assertThat(dependency.getArtifactId()).isEqualTo("spring-cloud-starter-bootstrap");
-  }
-
-  @Test
-  void shouldGetSpringCloudDependencies() {
-    Dependency dependency = SpringCloudConfig.springCloudDependencies();
-
-    assertThat(dependency.getGroupId()).isEqualTo("org.springframework.cloud");
-    assertThat(dependency.getArtifactId()).isEqualTo("spring-cloud-dependencies");
-  }
-
-  @Test
   void shouldGetSpringCloudConfigClient() {
     Dependency dependency = SpringCloudConfig.springCloudConfigClient();
